requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
python-dotenv==1.0.0

# Web scraping
selenium==4.16.0
beautifulsoup4==4.12.2

# AI and API integrations
openai==1.6.1
requests==2.31.0

# Image processing
Pillow==10.1.0
rembg==2.0.50

# Data handling
pandas==2.1.4
python-dateutil==2.8.2

# Utilities
python-slugify==8.0.1
validators==0.22.0

# Development
pytest==7.4.3
black==23.12.1